tls-sessions: Pass correct type for printf %u

+++ lib/tls/tls-sessions.c
@@ -34,7 +34,7 @@ lws_tls_session_tag_discrete(const char *vhname, const char *host,
 	 * different client certs.
 	 */
 
-	lws_snprintf(buf, len, "%s_%s_%u", vhname, host, port);
+	lws_snprintf(buf, len, "%s_%s_%u", vhname, host, (unsigned int)port);
 }
 
 int
